+++ app/helpers/admin/form_helper.rb
@@ -50,11 +50,12 @@ class Admin
     end
 
     def checkbox_tag_for(category)
+      cb_name = category.type == 'service' ? 'service[category_ids][]' : 'situation[category_ids][]'
       check_box_tag(
-        'service[category_ids][]',
-        category.id,
-        @taxonomy_ids.include?(category.taxonomy_id),
-        id: "category_#{category.taxonomy_id}"
+          cb_name,
+          category.id,
+          @taxonomy_ids.include?(category.taxonomy_id),
+          id: "category_#{category.taxonomy_id}"
       )
     end