+++ app/models/category.rb
@@ -2,6 +2,9 @@ class Category < ApplicationRecord
   self.inheritance_column = nil
   has_and_belongs_to_many :services
 
+  scope :services, -> { where(type: 'service') }
+  scope :situations, -> { where(type: 'situation') }
+
   validates :name, :taxonomy_id,
             presence: { message: I18n.t('errors.messages.blank_for_category') }