+++ db/migrate/20190304172224_add_type_to_categories.rb
@@ -2,7 +2,7 @@ class AddTypeToCategories < ActiveRecord::Migration[5.1]
   def up
     add_column :categories, :type, :string
     Category.update_all( {type: 'service'} )
-    change_column :categories, :type, :string, null: false
+    change_column :categories, :type, :string
   end
   def down
     remove_column :categories, :type, :string