+++ spec/features/admin/services/update_categories_spec.rb
@@ -3,10 +3,10 @@ require 'rails_helper'
 feature 'Update categories' do
   background do
     create_service
-    emergency = Category.create!(name: 'Emergency', taxonomy_id: '101')
-    emergency.children.create!(name: 'Disaster Response', taxonomy_id: '101-01')
-    emergency.children.create!(name: 'Subcategory 2', taxonomy_id: '101-02')
-    emergency.children.create!(name: 'Subcategory 3', taxonomy_id: '101-03')
+    emergency = Category.create!(name: 'Emergency', taxonomy_id: '101', type: 'service')
+    emergency.children.create!(name: 'Disaster Response', taxonomy_id: '101-01', type: 'service')
+    emergency.children.create!(name: 'Subcategory 2', taxonomy_id: '101-02', type: 'service')
+    emergency.children.create!(name: 'Subcategory 3', taxonomy_id: '101-03', type: 'service')
 
     login_super_admin
     visit '/admin/locations/vrs-services'